pyproject.toml
[tool.poetry]
name = "batemanrecords"
version = "0.1.0"
description = "automated patrick bateman walking to music uploads to instagram"
authors = ["codeblech <my.malikyash@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
spotipy = "^2.24.0"
pillow = "^11.0.0"
tqdm = "^4.66.5"
typer = "^0.12.5"
furo = "^2024.8.6"
sphinxcontrib-video = "^0.2.1"
yt-dlp = {extras = ["default"], version = "^2026.2.4"}
beautifulsoup4 = "^4.14.3"
lxml = "^6.0.2"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
sphinx = "^8.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
